move isNoncanonAndNoDepartures to the one place it's actually needed
