+++ apps/site/assets/ts/models/route-patterns.ts
@@ -1,9 +1,5 @@
 import { minBy, sortBy } from "lodash";
-import {
-  GroupedRoutePatterns,
-  RoutePatternWithPolyline
-} from "../stop/stop-redesign-loader";
-import { DepartureInfo } from "./departureInfo";
+import { GroupedRoutePatterns } from "../stop/stop-redesign-loader";
 
 type RoutePatternGroup = GroupedRoutePatterns[keyof GroupedRoutePatterns];
 export type RoutePatternGroupEntries = [
@@ -21,13 +17,5 @@ const sortedGroupedRoutePatterns = (
       minBy(routePatterns, "sort_order")?.sort_order
   );
 
-const isNoncanonicalAndNoDepartures = (
-  routePatterns: RoutePatternWithPolyline[],
-  departures: DepartureInfo[]
-): boolean => {
-  const isNonCanonical = !routePatterns.find(rp => !!rp.canonical);
-  return isNonCanonical && departures.length === 0;
-};
-
 // eslint-disable-next-line import/prefer-default-export
-export { sortedGroupedRoutePatterns, isNoncanonicalAndNoDepartures };
+export { sortedGroupedRoutePatterns };
